fix: prevent conflict between user id and commenter id
If the user id is the same as a commenter id, the function
get_avatar_url first retrieve the commenter instead of the user. So in
some cases, a logged in user had Wapuu as Gravatar instead of its own
Gravatar (Wapuu is the first comment on a fresh install of WP).
To fix this, I check if an user exists with that ID and use the user
object to retrieve the Gravatar.

Note: some conflicts may happen if an existing user has the same ID as
a guest commenter. I only tested on a small database so I don't see any
issue.